+++ CHANGELOG.md
@@ -16,12 +16,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
-- A sub-module skeleton: Gist
+- Sub-module gist
 
 ### Changed
-- Add xml version and encoding to default ocmod skeleton
+- Move default editor config to Alpha gist
 - Change method sequence for admin model
-- When saving sub-module get an error, only delete sub-module content
+- Delete sub-module content when there is an error saving
 
 ### Fixed
 - Properly handle invalid ocmod xml on sub-module save